+++ MKFoundationExample/MKFoundationExample/Presentation Layer/CheckBoxListScene/View/Cells/CheckBoxListCell.swift
@@ -11,7 +11,6 @@ class CheckBoxListCell: UITableViewCell {
     
     lazy var mkCheckBox: MKCheckBox = {
         let v = MKCheckBox()
-
         return v
     }()
