+++ Sources/MKfoundation/Component/CheckBox/Protoccol/MKCheckBoxDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  MKCheckBoxDelegate.swift
+//
+
+
+import Foundation
+
+public protocol MKCheckBoxDelegate: AnyObject {
+    
+    func didSelected(_ selectBox: MKCheckBox)
+    func didChangeStatus(_ selectBox: MKCheckBox, status: Bool)
+}
