--- Sources/MKfoundation/Component/MKCheckBox/Protoccol/MKCheckBoxDelegate.swift
@@ -1,12 +1,0 @@
-//
-//  MKCheckBoxDelegate.swift
-//
-
-
-import Foundation
-
-public protocol MKCheckBoxDelegate: AnyObject {
-    
-    func didSelected(_ selectBox: MKCheckBox)
-    func didChangeStatus(_ selectBox: MKCheckBox, status: Bool)
-}
