+++ Sources/MKfoundation/Component/SelectBox/Models/MKSelectBoxOptions.swift.swift
@@ -9,7 +9,7 @@ import UIKit
 /// 셀렉트 박스 를 구성하는 값을 적용합니다
 /// - Parameters:
 ///   - placeHolder: placeholer text
-///   - title: 텍스트 필드 제목 영역 입니다 적용 시 최대 높이가 변경됩니다
+///   - title: 필드 제목 영역 입니다 적용 시 최대 높이가 변경됩니다
 ///   - helperText: 하단 설명 영역입니다 적용 시 최대 높이가 변경됩니다
 
 public struct MKSelectBoxOptions {
