+++ Sources/MKfoundation/Component/TextView/View/MKTextView.swift
@@ -314,9 +314,13 @@ public extension MKTextView {
     /// - Parameters:
     ///   - option: `MKTextViewOptions` 를 참고하세요
     ///   - tableView: TableView Cell 안에 넣는 경우 적용합니다
-    ///   - important:TableView를 넘기지 않는 경우 에러 메시지로 인한 텍스트 높이 변경이 수동으로 이뤄져야 합니다
+    ///   - important: TableView를 넘기지 않는 경우 에러 메시지로 인한 텍스트 높이 변경이 수동으로 이뤄져야 합니다
     
-    func configure(option: MKTextViewOptions, tableView: UITableView? = nil, customToolbar: Bool = false) {
+    func configure(
+        option: MKTextViewOptions,
+        tableView: UITableView? = nil,
+        customToolbar: Bool = false
+    ) {
         self.tableView = tableView
         self.options = option
